local_printer/requirements.txt
pystray
Pillow
qrcode[pil]
PyPDF2
requests
